backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
jinja2==3.1.4
pydantic==2.7.4
email-validator==2.1.1
firebase-admin==6.5.0
flask
python-dotenv
jinja2